All classes and constraints exec in TclRAL for TWO subsystems
We loop through two subsystems, first creating all class relvars for both subsystems and then putting in all the relationship constraints. If a class is not present for any given relationship it is skipped. This is due to the imported class not be resolved yet (not enough subsystems read in).